TextToSql/requirements.txt
streamlit
google-generativeai
python-dotenv
tabulate
sqlparse